Move Timer class and chang docstrings
